// packages/braiins-client/src/generated/types.ts
export interface paths {
    "/spot/settings": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Retrieve market settings & rules */
        get: operations["spotGetMarketSettings"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/spot/stats": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get market statistics */
        get: operations["spotGetMarketStats"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/spot/orderbook": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** Get spot market order book snapshot */
        get: operations["spotGetOrderbookSnapshot"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        /** @enum {string} */
        SpotInstrumentStatus: "SPOT_INSTRUMENT_STATUS_UNSPECIFIED" | "SPOT_INSTRUMENT_STATUS_ACTIVE" | "SPOT_INSTRUMENT_STATUS_HALTED";
        MarketSettings: {
            status: components["schemas"]["SpotInstrumentStatus"];
            /** @description Order book decimalization step in satoshi. */
            tick_size_sat: components["schemas"]["Double"];
            /**
             * @description Market log10 multiplier for the hash rate. For example set it 18 to have EH/day as the base unit.
             * @example 18
             */
            hr_multiplier_log10: components["schemas"]["Int32"];
            /**
             * @description Human readable market log10 multiplier, e.g. 18 -> "EH/day", 17 -> "100PH/day"
             * @example EH/day
             * @example 100PH/day
             * @example 10PH/day
             */
            hr_unit: string;
            min_bid_price_sat: components["schemas"]["Double"];
            max_bid_price_sat: components["schemas"]["Double"];
            min_ask_price_sat: components["schemas"]["Double"];
            max_ask_price_sat: components["schemas"]["Double"];
            /** @description Minimum bid amount for bids with no speed limit in satoshi. */
            min_bid_amount_sat: components["schemas"]["Double"];
            /** @description Minimum bid amount for bids with no speed limit in satoshi. */
            max_bid_amount_sat: components["schemas"]["Double"];
            min_bid_speed_limit_ph: components["schemas"]["Double"];
            max_bid_speed_limit_ph: components["schemas"]["Double"];
            /** @description Max allowed idle time for any bid in seconds. For example 7 days = 604800. */
            max_bid_idle_time_s: components["schemas"]["Uint32"];
            /** Format: date-time */
            created: string;
            max_bids_per_subaccount: components["schemas"]["Uint32"];
            max_asks_per_subaccount: components["schemas"]["Uint32"];
            /** @description Minimum period required to be able to cancel a bid in seconds. */
            bid_grace_period_s: components["schemas"]["Uint32"];
            /** @description Minimum period required to be able to cancel an ask in seconds. */
            ask_grace_period_s: components["schemas"]["Uint32"];
            /** @description Price is allowed to be decreased only when this period passes since last decrease (in seconds). */
            min_bid_price_decrease_period_s: components["schemas"]["Uint32"];
            /** @description Speed limit is allowed to be decreased only when this period passes since last decrease (in seconds). */
            min_bid_speed_limit_decrease_period_s: components["schemas"]["Uint32"];
            /** @description Minimum bid amount for bids with speed limit in satoshi. */
            min_limited_bid_amount_sat: components["schemas"]["Double"];
            /** @description Maximum bid amount for bids with speed limit in satoshi. */
            max_limited_bid_amount_sat: components["schemas"]["Double"];
            /** @description Minimum duration for bids with speed limit in seconds. */
            min_limited_bid_duration_s: components["schemas"]["Uint32"];
        };
        ProfileIdentifier: {
            profile_name: string;
            client_name: string;
        };
        SpotGetBidsResponse: {
            items: components["schemas"]["SpotGetBidsResponseItem"][];
        };
        SpotGetBidsResponseItem: {
            bid: components["schemas"]["SpotMarketBid"];
            counters_estimate: components["schemas"]["SpotMarketBidCounters"];
            counters_committed: components["schemas"]["SpotMarketBidCounters"];
            state_estimate: components["schemas"]["SpotMarketBidState"];
            /** @description Last network failure (if any). */
            last_network_failure?: components["schemas"]["UpstreamFailure"];
        };
        SpotGetBidDetailResponse: {
            /** @description Complete bid information including all metadata */
            bid: components["schemas"]["SpotMarketBid"];
            /** @description Estimated counters for the bid (may be ahead of committed values) */
            counters_estimate: components["schemas"]["SpotMarketBidCounters"];
            /** @description Committed counters for the bid (confirmed values) */
            counters_committed: components["schemas"]["SpotMarketBidCounters"];
            /** @description Current estimated state of the bid including speed and progress */
            state_estimate: components["schemas"]["SpotMarketBidState"];
            /** @description Last network failure related to the bid, if any. For bids with external destinations only. */
            last_network_failure?: components["schemas"]["UpstreamFailure"];
            /** @description History of bid status changes and updates */
            history?: components["schemas"]["SpotBidHistoryItem"][];
        };
// ...
        SpotPlaceBidRequest: {
            /** @description Client assigned ID of this bid, optional. */
            cl_order_id?: string;
            /** @description Upstream specification for orders of dest type UPSTREAM */
            dest_upstream: components["schemas"]["UpstreamSpecification"];
            /** @description Optional speed limit in PH/s. */
            speed_limit_ph?: components["schemas"]["Double"];
            /** @description Order amount in satoshi. */
            amount_sat: components["schemas"]["Double"];
            /** @description Bid price specified in satoshi. */
            price_sat: components["schemas"]["Double"];
            /** @description Remark (optional). */
            memo?: string;
        };
        PlaceOrderResponse: {
            /** @description Autogenerated order ID. */
            id: string;
            /** @description Client order ID passed through. */
            cl_order_id: string;
        };
        SpotEditBidRequest: {
            /** @description Bid ID (exclusive with cl_order_id). */
            bid_id?: string;
            /** @description Client assigned ID of this bid for identification (exclusive with bid_id). */
            cl_order_id?: string;
            /** @description New order amount (total). If provided, it must be greater than previous amount value. */
            new_amount_sat?: components["schemas"]["Double"];
            /** @description New bid price in satoshi. */
            new_price_sat?: components["schemas"]["Double"];
            /** @description New speed limit in PH/s. Set to 0 to disable. If not provided, speed limit is not changed. */
            new_speed_limit_ph?: components["schemas"]["OptionalDouble"];
            /** @description Remark (optional). */
            memo?: string;
        };
        /** @description IDs affected by the operation (canceled successfully). */
        CancelResponse: {
            affected_ids: components["schemas"]["MultipleStringIds"];
        };
        MultipleStringIds: {
            /** @description Client order IDs for identification. */
            id: string[];
        };
        GetFeeStructureResponse: {
            spot_fees: components["schemas"]["SpotFeePublic"][];
        };
// ...
        GetAccountBalancesResponse: {
            accounts: components["schemas"]["AccountBalance"][];
        };
        AccountBalance: {
            /** @description Related subaccount */
            subaccount: string;
            /**
             * @description Account currency.
             * @example USDC
             */
            currency: string;
            /** @description Actual total account balance (including blocked funds). */
            total_balance_sat: components["schemas"]["Int64"];
            /** @description Account balance minus blocked funds. Funds available for withdrawals / purchases. */
            available_balance_sat: components["schemas"]["Int64"];
            /** @description Blocked amount (in orders, etc...). */
            blocked_balance_sat: components["schemas"]["Int64"];
            /** @description Total deposited. */
            total_deposited_sat: components["schemas"]["Int64"];
            /** @description Total withdrawn from this account. */
            total_withdrawn_sat: components["schemas"]["Int64"];
            /** @description Total spent on spot market bids (net). */
            total_spot_spent_sat: components["schemas"]["Int64"];
            /** @description Total revenue from spot market asks (gross). */
            total_spot_revenue_gross_sat: components["schemas"]["Int64"];
            /** @description Total revenue from spot market asks (net). */
            total_spot_revenue_net_sat: components["schemas"]["Int64"];
            /** @description Total spent on spot fees (buy) */
            total_spent_spot_buy_fees_sat: components["schemas"]["Int64"];
            /** @description Total spent in fees. */
            total_spent_spot_sell_fees_sat: components["schemas"]["Int64"];
            total_spent_fees_sat: components["schemas"]["Int64"];
            /** @description True if there is a pending withdrawal (clients may only withdraw the full amount). */
            has_pending_withdrawal: boolean;
        };
        GetTransactionsResponse: {
            transactions: components["schemas"]["Transaction"][];
        };
        GetOnChainTransactionsResponse: {
            transactions: components["schemas"]["OnChainTransaction"][];
        };
        Transaction: {
            /** @description Transaction type (deposit, withdrawal, fee, etc...). */
            tx_type: string;
            /** @description Order amount in satoshi. */
            amount_sat: components["schemas"]["Double"];
            /** @description Additional details. */
            details: string;
            /**
             * Format: date-time
             * @description Transaction timestamp.
             */
            timestamp: string;
        };
// ...
        SpotGetMarketStatsResponse: {
            status: components["schemas"]["SpotInstrumentStatus"];
            /** @description Shares sold here in last 24 hours. In millions of shares. */
            volume_24h_m: components["schemas"]["Double"];
            best_bid_sat: components["schemas"]["Double"];
            best_ask_sat: components["schemas"]["Double"];
            /** @description Last second's average price matched. In satoshi. */
            last_avg_price_sat: components["schemas"]["Double"];
            /** @description Hash rate being matched, estimate on 10m rolling window. In PH/s. */
            hash_rate_matched_10m_ph: components["schemas"]["Double"];
            /** @description Hash rate available in asks, estimate on 10m rolling window. In PH/s. */
            hash_rate_available_10m_ph: components["schemas"]["Double"];
        };
        SpotGetOrderBookResponse: {
            /** @description List of bid items. */
            bids: components["schemas"]["BidItem"][];
            /** @description List of ask items. */
            asks: components["schemas"]["AskItem"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}